Move varargs tests into existing class
